fix(build): Ensure that main jar does not contain test classes.
Modify maven build to produce the JARs more clean (no test stuff in main
jar, junit4/junit5 jars contain shared stuff and specific stuff).

Adapt other project builds accordingly.

Moved some test classes to shared project folder. Moved some classes to
"test.mock" to "mock" package to make build more easy and clearly mark
them as production code.

fixes #819
