+++ engine-plugins/spin-plugin/src/test/java/org/operaton/spin/plugin/script/SpinJRubyScriptTaskTest.java
@@ -16,36 +16,28 @@
  */
 package org.operaton.spin.plugin.script;
 
+import static org.assertj.core.api.Assertions.assertThat;
+
+import org.junit.jupiter.api.Disabled;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.RegisterExtension;
 import org.operaton.bpm.engine.RepositoryService;
 import org.operaton.bpm.engine.RuntimeService;
 import org.operaton.bpm.engine.repository.Deployment;
 import org.operaton.bpm.engine.runtime.ProcessInstance;
-import org.operaton.bpm.engine.test.ProcessEngineRule;
+import org.operaton.bpm.engine.test.junit5.ProcessEngineExtension;
 import org.operaton.bpm.engine.variable.VariableMap;
 import org.operaton.bpm.engine.variable.Variables;
 import org.operaton.bpm.model.bpmn.Bpmn;
 import org.operaton.bpm.model.bpmn.BpmnModelInstance;
 
-import org.junit.Rule;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Disabled;
-import org.junit.jupiter.api.Test;
-
-import static org.assertj.core.api.Assertions.assertThat;
-
 class SpinJRubyScriptTaskTest {
 
-  @Rule
-  public ProcessEngineRule engineRule = new ProcessEngineRule();
-
-  private RuntimeService runtimeService;
-  private RepositoryService repositoryService;
-
-  @BeforeEach
-  void setUp() {
-    this.runtimeService = engineRule.getRuntimeService();
-    this.repositoryService = engineRule.getRepositoryService();
-  }
+  @RegisterExtension
+  static ProcessEngineExtension engineRule = ProcessEngineExtension.builder().build();
+  
+  RuntimeService runtimeService;
+  RepositoryService repositoryService;
 
   @Test
   @Disabled("CAM-11114")
