+++ engine-spring/src/main/java/org/operaton/bpm/engine/spring/SpringExpressionManager.java
@@ -22,7 +22,7 @@ import org.operaton.bpm.engine.impl.el.JuelExpressionManager;
 import org.operaton.bpm.engine.impl.el.ReadOnlyMapELResolver;
 import org.operaton.bpm.engine.impl.el.VariableContextElResolver;
 import org.operaton.bpm.engine.impl.el.VariableScopeElResolver;
-import org.operaton.bpm.engine.test.mock.MockElResolver;
+import org.operaton.bpm.engine.impl.mock.MockElResolver;
 import jakarta.el.ArrayELResolver;
 import jakarta.el.BeanELResolver;
 import jakarta.el.CompositeELResolver;
