+++ spring-boot-starter/starter-test/src/main/java/org/operaton/bpm/spring/boot/starter/test/helper/StandaloneInMemoryTestConfiguration.java
@@ -19,8 +19,8 @@ package org.operaton.bpm.spring.boot.starter.test.helper;
 import org.operaton.bpm.engine.impl.cfg.ProcessEnginePlugin;
 import org.operaton.bpm.engine.impl.cfg.StandaloneInMemProcessEngineConfiguration;
 import org.operaton.bpm.engine.impl.history.HistoryLevel;
+import org.operaton.bpm.engine.impl.mock.MockExpressionManager;
 import org.operaton.bpm.engine.test.junit5.ProcessEngineExtension;
-import org.operaton.bpm.engine.test.mock.MockExpressionManager;
 
 import java.util.Arrays;
 import java.util.List;
