+++ test-utils/assert/core/src/test/java/org/operaton/bpm/engine/test/assertions/bpmn/ProcessEngineTestsJobTest.java
@@ -17,10 +17,11 @@
 package org.operaton.bpm.engine.test.assertions.bpmn;
 
 import org.operaton.bpm.engine.ProcessEngineException;
+import org.operaton.bpm.engine.impl.mock.Mocks;
 import org.operaton.bpm.engine.runtime.ProcessInstance;
 import org.operaton.bpm.engine.test.Deployment;
 import org.operaton.bpm.engine.test.assertions.helpers.ProcessAssertTestCase;
-import org.operaton.bpm.engine.test.mock.Mocks;
+
 import static org.operaton.bpm.engine.test.assertions.bpmn.BpmnAwareTests.*;
 
 import org.junit.jupiter.api.Test;
