+++ test-utils/assert/core/src/test/java/org/operaton/bpm/engine/test/assertions/bpmn/ProcessInstanceAssertJobTest.java
@@ -17,11 +17,12 @@
 package org.operaton.bpm.engine.test.assertions.bpmn;
 
 import org.operaton.bpm.engine.ProcessEngineException;
+import org.operaton.bpm.engine.impl.mock.Mocks;
 import org.operaton.bpm.engine.runtime.JobQuery;
 import org.operaton.bpm.engine.runtime.ProcessInstance;
 import org.operaton.bpm.engine.test.Deployment;
 import org.operaton.bpm.engine.test.assertions.helpers.ProcessAssertTestCase;
-import org.operaton.bpm.engine.test.mock.Mocks;
+
 import static org.operaton.bpm.engine.test.assertions.bpmn.BpmnAwareTests.assertThat;
 import static org.operaton.bpm.engine.test.assertions.bpmn.BpmnAwareTests.execute;
 import static org.operaton.bpm.engine.test.assertions.bpmn.BpmnAwareTests.jobQuery;
